# main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a document for Tainosystems",
        "termsOfService": "http://tainosystems.com",
        "contact": {
            "name": "Peterson Chery",
            "email": "Petersonchery33@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.10.146:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/clients": {
            "post": {
                "description": "Add a client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Add a client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "project",
                        "name": "Client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Client"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/clients/sup/{id}": {
            "delete": {
                "description": "Delete a Client By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Delete a Client By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID client",
                        "name": "ID_client",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Client"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/clients/up/{id}": {
            "put": {
                "description": "Update a Client By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "UpdateClient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID client",
                        "name": "ID_client",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "project",
                        "name": "Client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Client"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/clients/{id}": {
            "get": {
                "description": "Get a client by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Get a client by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID client",
                        "name": "ID_client",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Client"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/con": {
            "post": {
                "description": "Connexion pour un utilisateur",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connexion"
                ],
                "summary": "Connexion pour un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "project",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Users"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Users"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Client": {
            "type": "object",
            "properties": {
                "email_client": {
                    "type": "string"
                },
                "id_client": {
                    "type": "integer"
                },
                "nom_client": {
                    "type": "string"
                },
                "password_client": {
                    "type": "string"
                }
            }
        },
...
        "model.Users": {
            "type": "object",
            "properties": {
                "email_user": {
                    "description": "Id_role   int //on va avoir 3 roles: admin, bibliothecaire, simpleUser\n\tPermission []string // les permissions sont les actions que les roles peuvent effectuer",
                    "type": "string"
                },
                "id_role": {
                    "type": "integer"
                },
                "id_session": {
                    "type": "integer"
                },
                "id_tenant": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "nom_user": {
                    "type": "string"
                },
                "password_user": {
                    "type": "string"
                }
            }
        }
    }
}